reorganize sketches and stop using type variables for now
