+++ test/test_infer_shapes.py
@@ -41,12 +41,44 @@ def compute_sketches(cs:Dict[str,List[str]], callgraph:Dict[str,List[str]]):
 class InferTypesTest(unittest.TestCase):
 
     def test_input_arg_capability(self):
-        constraints = {"f":["f.in_0 ⊑ A","A.load.σ1@0 ⊑ B"]}
-        callgraph = {"f":[]}
+        """
+        f.in_0 should get the load.σ1@0 capability even if
+        we don't know anything about its type.
+        The same with g.out
+        """
+        constraints = {
+            # T-inheritR
+            "f":["f.in_0 <= A","A.load.σ1@0 <= B","B.load.σ4@4 <= C"],
+            # T-inheritL
+            "g":["A <= g.out","A.load.σ1@0 <= B","B.load.σ4@4 <= C"]
+        }
+        callgraph = {"f":[],"g":[]}
         (gen_cs, sketches) = compute_sketches(constraints,callgraph)
 
         f_sketch = sketches[DerivedTypeVariable("f")]
         assert SchemaParser.parse_variable("f.in_0") in f_sketch.lookup
+        assert SchemaParser.parse_variable("f.in_0.load.σ1@0.load.σ4@4") in f_sketch.lookup
 
+        f_sketch = sketches[DerivedTypeVariable("g")]
+        assert SchemaParser.parse_variable("g.out") in f_sketch.lookup
+        assert SchemaParser.parse_variable("g.out.load.σ1@0") in f_sketch.lookup
+        assert SchemaParser.parse_variable("g.out.load.σ1@0.load.σ4@4") in f_sketch.lookup
+
+    def test_input_arg_capability_transitive(self):
+        """
+        g.in_0 should get the load.σ1@0 capability from f.
+        """
+        constraints = {
+            "f":["f.in_0 <= A","A.load.σ1@0 <= B","B.load.σ4@4 <= C"],
+            "g":["g.in_0 <= C", "C <= f.in_0", "C <= g.out" ]
+        }
+        callgraph = {"g":["f"]}
+        (gen_cs, sketches) = compute_sketches(constraints,callgraph)
+
+        f_sketch = sketches[DerivedTypeVariable("g")]
+        assert SchemaParser.parse_variable("g.in_0") in f_sketch.lookup
+        assert SchemaParser.parse_variable("g.in_0.load.σ1@0.load.σ4@4") in f_sketch.lookup
+        assert SchemaParser.parse_variable("g.out.load.σ1@0.load.σ4@4") in f_sketch.lookup
+
